Add server.py (#5)
* Add firebase streams

Create 3 streams for 'currentMatchNumber', 'cycleNumber',
'TempTeamInMatchDatas'. Streams are created by create_streams() and
returned as a dict. create_streams() is designed to be able to only
create specific streams in the event that streams need to be restarted
(TODO).

* Add infinite while loop and automatic stream restart

Infinite while loop handles stream restart + starts calculations when
new tempTIMDs are available.

* Add handling for CTRL+C

Running streams to firebase prevents normal CTRL+C from exiting the
script. In order to exit, streams must be closed.

* Add create_file_path() to use repository directory

This allows for other paths to use a path relative to inside the
repository folder.

* Delete cached data when TEMP_TIMD_STREAM is (re)started.

This is to remove any outdated data, since all data is re-downloaded
when the TEMP_TIMD_STREAM is restarted.

* Add lastServerRun to monitor if the server is offline

Each cycle of the while loop updates a key on firebase storing the epoch
time that the run completed.  Used to monitor if the server is offline
by checking if an excess amount of time has passed since the last run.

* Add comment to explain restarting streams

* Add data verification for integer database streams

Validates that data was correctly received and in its expected format by
checking the event type, the path (to ensure it is at the correct
location), and that the data type is an integer.  This is to prevent
crashes in the event of an extreme database error or if the variable is
non-existent.

* Fix typo in comment in server.py

* Update server.py to import firebase_communicator

firebaseCommunicator was renamed to firebase_communicator to comply with
PEP8.

* Add check for 'data' directory before deleting it

Prevents error from being thrown if directory is missing.

* Add handling for tempTIMD stream in server.py

Saves tempTIMDs in individual .txt files.  Removes files when data is
deleted.  Validates the correct number of children in tempTIMDs.