Add tempTIMD timeline decompression (#1)
* Decompress tempTIMD timeline

* Fixes the problem with the climb dictionaries.

Changes code for dealing with the climb dictionaries. Removes the code for
splitting the data at the colens. Changes variable names for the decompression
dictionaries. Fixes formatting for the doc string.

* Updates the dictionaries used for decompression.

Adds new data points into the compression keys and values.

* Fixes the line length formatting.

Shortens the comments and the lines of code so they are not over the limit.

* Adds code to deal with no show teams.

Adds an "if not" statement to catch if a team does not have any timeline data.

* Fixes the length of the lines.

Shortens the length of the comments and the lenth of lines of code.

* Changes name of function and argument.

Changes to remove capitals from function names so they conform to snake-case
style.

* Fixes the merge conflict fixes that were not saved.

Removes the code for the merge conflicts that was push to git hub.

* Updates the comments.

Updates comments with periods and new function name.

* Adds return statement to the function.

* Adds decompress_temp_timd function.

Function calls timeline and header functions. Takes in the compressed
tempTIMD and outputs a dictionary of the decompressed tempTIMD.

* Removes code concerning the underscore.

Removes the code checking for the underscore and the split at the underscore.

* Fixes syntax error for decompress_temp_timd function.

Rearanges the function so it returns the full decompressed temp timd.

* Updates file doc string.

* Updates and fixes problems in doc strings.

* Fixes doc strings.

* Changes variable names.

* Fixed formatting in doc string for tempTIMD timeline decompression.

* Changes variable name for splitting by commas.

* Changes variable name from char to character.

* Changes comments to make more clear.

* Moves decompression dictionaries outside of function.

* Modify tempTIMD timeline compression (#4)

* Update docstrings in decompress.py

* Rename 'action_dic' to 'decompressed_action'

* Separate one-line statement into three lines

* Simplify climb if-statement

* Rename 'climb_dic' to 'decompressed_climb'

* Change tuple in if-statement to list

* Separate one-line statement into three lines

* Simplify climb decompression

* Fix continuned indentation

* Update comments in decompress.py

* Fixes mistake in TEMP_TIMD_COMPRESSION_KEYS dictionary.

* Add int type converting + re-structure logic (#6)

* Use 'isalpha()' instead of 'ascii_letters'

* Add int type converting + re-structure logic

* Debug tempTIMD timeline decompression

Add handling for trailing comma
Add handling for the first key of an action
Save a key's index and value after saving the previous key:value pair
Remove curly brackets before iterating through climb